requirements.txt
pandas
torch
transformers
streamlit
plotly
pydantic
pydantic-settings
PyYAML
emoji
scikit-learn
pytest
